Add union decoding:
This commit introduces the Union interface, which types may implement
to specify to the system that it represents an xdr union.  By
implementing `SwitchFieldName` the union is specifying where the
discriminant should be read/written.  `ArmForSwitch` returns the name
of the data field to read/write for the union’s value.