Check xdrmaxsize tag in Decoder.decodeUnion
decodeUnion method of Decoder wasn't checking for tag xdrmaxsize in the
fields of an union. Attacker could exploit this by creating an union
with a very long string (large XDR length value) and passing it to
Decoder. By doing this attacker could allocate a lot of memory and crash
an app that's using go-xdr.

This commit fixes this by checking for xdrmaxsize tag in decodeUnion.