Update README and LICENSE (#6)
Update the README to remove the statement that this is a temporary fork, remove details about `xdr2` and instead mention that `xdr3` is the latest version. Also add the Stellar Development Foundation as an additional copyright holder since 2015.

Why:

When the `go-xdr` project was originally forked from `davecgh/go-xdr` to `nullstyle/go-xdr` and then to `stellar/go-xdr` there was an assumption we'd take our improvements upstream. But our version has continued to fork significantly and now implements features in ways that are different with other open PRs on the upstream. An issue was opened at davecgh/go-xdr#17 to explore what options there were to merge changes upstream but the owner of the repository did not respond. While another member of the repository responded, the effort to merge upstream is somewhat significant so I think we'd only want to do that if we got strong buy-in from the repository owner.

Given that our version has diverged significantly and is used in Stellar's encoding and decoding of XDR in Go projects we're going to continue developing it here. We already took the first step of doing that by accepting the community contribution #3/#5.

I've added SDF to the LICENSE from 2015 because in this repository the changes made by SDF employees on both the `nullstyle/go-xdr` and `stellar/go-xdr` repositories occurred starting in the year 2015.

For #4