Use GitHub Actions instead of TravisCI (#9)
Move the CI checks from TravisCI to GitHub Actions, and have all checks run on all code in the repository, and fix some issues that resulted from running the checks on all code. Also added a simple `go.mod` file to support those changes.

We use CircleCI for all our builds at SDF and initially I explored in #8 just porting this to CircleCI too, but I think it's worth us experimenting with GitHub Actions as it's very powerful and available to all individuals on GitHub and for open source libraries like go-xdr we do not need to grant others access to run their PRs on our CircleCI account.

Some of the changes made in this PR are because checks were only running on one the `xdr2` package, and this makes them run on all packages. I simplified the existing `goclean.sh` script and renamed it to `gotest.sh`.

For #4