Fix decoding nil pointer to the previously used object (#13)
This commit fixes a bug in `Decoder.decodePtr`. When decoding a struct
with a pointer field into previously used object (ex. in a loop with
object created outside of the loop) the previous non-`nil` pointer value
was not overwritten if the new value was `nil`. See
`TestDecodeNilPointerIntoExistingObjectWithNotNilPointer` for more
context.