Make encoder/decoder underlying errors unwrappable (#15)
Add `Unwrap` methods to the `MarshalError` and `UnmarshalError` types that return the underlying IO errors if present.

The `MarshalError` and `UnmarshalError` error types wrap IO and other errors. Since Go 1.13 if an error provides an `Unwrap` function that returns the underlying error, the `errors` stdlib package can be used to test if the error or any of its underlying errors are an error. This is helpful for testing for general underlying IO errors in code the uses the encoder/decoder as a reader.